Made survive-mode last longer to handle high speed
With the new gear ratio, the robot sometimes goes over the edge of the
arena with the front sensors before going into survival mode. Because of
the short duration of the survival mode, the robot is not able to break
and alter direction for long enough to back up to the arena edge.

This caused the robot to halt briefly after crossing the line, before
continuing on its way out of the arena.